Refactor group repository and jpa-dao
