+++ dao/src/main/java/com/hashmapinc/server/dao/sql/customergroup/CustomerGroupRepository.java
@@ -15,10 +15,8 @@
  */
 package com.hashmapinc.server.dao.sql.customergroup;
 
-import com.hashmapinc.server.common.data.CustomerGroup;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import com.hashmapinc.server.dao.model.nosql.CustomerGroupEntity;
+import com.hashmapinc.server.dao.model.sql.CustomerGroupEntity;
 import com.hashmapinc.server.dao.util.SqlDao;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
@@ -38,7 +36,7 @@ public interface CustomerGroupRepository  extends CrudRepository<CustomerGroupEn
                                                           @Param("idOffset") String idOffset,
                                                           Pageable pageable);
 
-    CustomerGroupEntity findByTenantIdAndCustomerIdAndTitle(String tenantId, String customerId, String title);
+    CustomerGroupEntity findByTenantIdAndCustomerIdAndTitle(String tenantId , String customerId , String title);
 
-    List<CustomerGroup> findByIdIn(List<String> id, PageRequest pageRequest);
+    List<CustomerGroupEntity> findByIdIn(List<String> id, Pageable pageable);
 }
